requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Cython==0.29.13
DBUtils==1.3
Django==2.1.10
django-cors-headers==3.0.2
django-crontab==0.7.1
django-filter==2.1.0
django-stubs==0.12.1
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.8
idna==2.8
isort==4.3.21
joblib==0.13.2
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
numpy==1.16.4
pandas==0.24.2
pep8==1.7.1
Pillow==6.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pymssql==2.1.4
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.3.5
typing-extensions==3.7.4
urllib3==1.25.3
wrapt==1.11.2
